move test_encode_linear_ge_order_encoding_native to order::
